[ feat ] 게시판 기능 기능 분리
1. 서버 요청을 담당하는 컴포넌트(postService)에 예외처리를 추가하고 사용하는 컴포넌트에서는 로직만 수행하게 변경함
