+++ next-researchlab/src/app/strategylab/[postNo]/page.tsx
@@ -1,35 +1,15 @@
 "use client";
 
-import { Post } from "@/types/types";
 import { useParams } from "next/navigation";
-import { useEffect, useState } from "react";
 
 import PostDetail from "@/components/strategyLab/postDetail";
-import postService from "@/services/postService";
+import useFetchPostDetail from "@/hooks/posts/useFetchPostDetail";
 
 const StrategyPostDetail = () => {
   const params = useParams();
   const postNo = typeof params?.postNo === "string" ? params.postNo : "";
-  const [isLoading, setIsLoading] = useState(true);
-  const [post, setPost] = useState<Post | null>(null);
 
-  useEffect(() => {
-    if (postNo) {
-      const fetchPost = async () => {
-        console.log("상세글 조회");
-
-        try {
-          const resPost = await postService.fetchPostDetail(postNo);
-          setPost(resPost);
-          setIsLoading(false);
-        } catch (error) {
-          console.log(error);
-        }
-      };
-
-      fetchPost();
-    }
-  }, [postNo]);
+  const { post, isLoading } = useFetchPostDetail(postNo);
 
   return (
     <div style={{ textAlign: "center", padding: "20px", color: "#e1eeeb" }}>
