+++ next-researchlab/src/services/postService.tsx
@@ -3,10 +3,15 @@ import axios from "axios";
 
 const postService = {
   async fetchAllPosts(page: number, size: number): Promise<Post[]> {
-    const res = await axios.get("/api/strategylab/posts", {
-      params: { page, size },
-    });
-    return res.data.content;
+    try {
+      const res = await axios.get("/api/strategylab/posts", {
+        params: { page, size },
+      });
+      return res.data.content;
+    } catch (e) {
+      console.log("전체 게시글 조회 실패", e);
+      return [];
+    }
   },
 
   async fetchMyPosts(
@@ -14,37 +19,66 @@ const postService = {
     page: number,
     size: number
   ): Promise<Post[]> {
-    const res = await axios.get(`/api/strategylab/members/${memberId}/posts`, {
-      params: { page, size },
-    });
-    return res.data.content;
+    try {
+      const res = await axios.get(
+        `/api/strategylab/members/${memberId}/posts`,
+        {
+          params: { page, size },
+        }
+      );
+
+      return res.data.content;
+    } catch (e) {
+      console.log("내 게시글 조회 실패", e);
+      return [];
+    }
   },
 
-  async fetchPostDetail(postNo: string): Promise<Post> {
-    const res = await axios.get(`/api/strategylab/posts/${postNo}`);
-    return res.data;
+  async fetchPostDetail(postNo: string): Promise<Post | null> {
+    try {
+      const res = await axios.get(`/api/strategylab/posts/${postNo}`);
+      return res.data;
+    } catch (e) {
+      console.log("게시글 상세조회 실패", e);
+      return null;
+    }
   },
 
   async createPost(formData: FormData) {
-    return await axios.post("/api/strategylab/posts", formData, {
-      headers: {
-        "Content-Type": "multipart/form-data",
-      },
-    });
+    try {
+      return await axios.post("/api/strategylab/posts", formData, {
+        headers: {
+          "Content-Type": "multipart/form-data",
+        },
+      });
+    } catch (e) {
+      console.log("글 작성 실패", e);
+      alert("게시글 작성에 실패하였습니다.");
+    }
   },
 
   async updatePost(postNo: number, formData: FormData) {
-    return await axios.put(`/api/strategylab/posts/${postNo}`, formData, {
-      headers: {
-        "Content-Type": "multipart/formData",
-      },
-    });
+    try {
+      return await axios.put(`/api/strategylab/posts/${postNo}`, formData, {
+        headers: {
+          "Content-Type": "multipart/formData",
+        },
+      });
+    } catch (e) {
+      console.log("게시글 수정 실패", e);
+      alert("게시글 수정에 실패하였습니다.");
+    }
   },
 
   async deletePost(postNo: number, memberId: string) {
-    return await axios.delete(`/api/strategylab/posts/${postNo}`, {
-      params: { memberId },
-    });
+    try {
+      return await axios.delete(`/api/strategylab/posts/${postNo}`, {
+        params: { memberId },
+      });
+    } catch (e) {
+      console.log("게시글 삭제 실패", e);
+      alert("게시글 삭제에 실패하였습니다.");
+    }
   },
 };
